requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.1
faiss-cpu==1.11.0.post1
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.73
langchain-text-splitters==0.3.9
langchainhub==0.1.21
langdetect==1.0.9
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.13
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
neo4j==5.28.2
networkx==3.5
nltk==3.9.1
numpy==2.3.2
olefile==0.47
openai==1.99.2
orjson==3.11.1
ormsgpack==1.10.0
packaging==24.2
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==5.9.0
pytest==8.4.1
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.6.1
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
unstructured==0.18.11
unstructured-client==0.42.2
urllib3==2.5.0
webencodings==0.5.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
pydantic>=2.6.4
python-dotenv>=1.0.1
pdfplumber>=0.10.3
tiktoken>=0.6.0
langchain-core>=0.1.50
langchain-openai>=0.1.6
openai>=1.30.1
langchain-community>=0.2.0  # optional but safe
langchain
openai
requests
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-groq
langchain-openai
pytesseract>=0.3.10
Pillow>=10.0.0
python-dotenv>=1.0.0
langgraph>=0.0.36
openai>=1.10.0
groq>=0.1.1
